+++ server/.temp/server_test.dart
@@ -4,7 +4,9 @@ import 'package:server/common/logger.dart';
 
 void main() async {
   var logger = Logger();
-  logger.logs.forEach((le) => print('${le.timestamp}\t${le.message}\n${le.level}\n\n'));
+  logger.logs.forEach(
+    (le) => print('${le.timestamp}\t${le.message}\n${le.source}\n${le.level}\n\n'),
+  );
 
   var handlerServer = HandlerService();
   handlerServer.initialize();